requirements/base.txt
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.7.0
redis==5.2.1
celery==5.4.0
boto3==1.35.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
